+++ templates/index.php
@@ -9,7 +9,8 @@ script('news', [
     'vendor/angular-route/angular-route.min',
     'vendor/angular-sanitize/angular-sanitize.min',
     'vendor/momentjs/min/moment-with-locales.min',
-    'build/app.min'
+    'vendor/masonry/dist/masonry.pkgd.min',
+    'build/app.min',
 ]);
 
 style('news', 'news.min');